Added Hare method, Ultra high speed evaluation and CMD user interface.
The old implemenation only supported Droop quota calculation. Now the quata calculation can be set in the configuration and provides the Hare method as well.

The delay in the UI which is nice in visualization/presenation mode can now set completely to off (slider completely to the right).

A CMD ui has been added